+++ bee-node/src/main.rs
@@ -1,9 +1,11 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+use bee_gossip::Keypair;
 use bee_node::{
-    plugins, print_banner_and_version, tools, ClArgs, EntryNodeBuilder, FullNodeBuilder, FullNodeConfig, NodeConfig,
-    NodeConfigBuilder, NodeStorageBackend,
+    plugins, print_banner_and_version, read_keypair_from_pem_file, tools, write_keypair_to_pem_file, ClArgs,
+    EntryNodeBuilder, EntryNodeConfig, FullNodeBuilder, FullNodeConfig, Local, NodeConfig, NodeConfigBuilder,
+    NodeStorageBackend, PemFileError,
 };
 use bee_runtime::node::NodeBuilder as _;
 
@@ -12,9 +14,12 @@ use bee_storage_rocksdb::storage::Storage;
 #[cfg(all(feature = "sled", not(feature = "rocksdb")))]
 use bee_storage_sled::storage::Storage;
 
+use log::warn;
+
 use std::{error::Error, path::Path};
 
 const CONFIG_PATH: &str = "./config.toml";
+const IDENTITY_PATH: &str = "./identity.pem";
 
 /// The entry point of the node software.
 #[tokio::main(flavor = "multi_thread")]
@@ -36,31 +41,54 @@ async fn main() -> Result<(), Box<dyn Error>> {
     print_banner_and_version(true);
 
     // Deserialize the config.
-    let config = deserialize_config(cl_args);
+    let identity_path = cl_args.identity_path().unwrap_or(Path::new(IDENTITY_PATH)).to_owned();
+    let (alias, has_identity_field, config) = deserialize_config(cl_args);
 
     // Initialize the logger.
     let logger_cfg = config.logger_config().clone();
     fern_logger::logger_init(logger_cfg)?;
 
+    // Establish identity.
+
+    // The identity used to be stored in the config file.
+    if has_identity_field {
+        warn!("The config file contains an identity field which will be ignored.");
+    }
+
+    let keypair = match read_keypair_from_pem_file(&identity_path) {
+        Ok(keypair) => keypair,
+        Err(PemFileError::FileRead(_)) => {
+            let keypair = Keypair::generate();
+            if let Err(e) = write_keypair_to_pem_file(identity_path, &keypair) {
+                panic!("Failed to write PEM file: {}", e);
+            };
+            keypair
+        }
+        Err(e) => panic!("Failed to establish identity: {}", e),
+    };
+
+    let local = Local::from_keypair(keypair, alias);
+
     // Start running the node.
     if config.run_as_entry_node() {
-        start_entrynode(config).await;
+        start_entrynode(local, config).await;
     } else {
-        start_fullnode(config).await;
+        start_fullnode(local, config).await;
     }
 
     Ok(())
 }
 
-fn deserialize_config(cl_args: ClArgs) -> NodeConfig<Storage> {
-    match NodeConfigBuilder::<Storage>::from_file(cl_args.config_path().unwrap_or_else(|| Path::new(CONFIG_PATH))) {
+fn deserialize_config(cl_args: ClArgs) -> (Option<String>, bool, NodeConfig<Storage>) {
+    match NodeConfigBuilder::<Storage>::from_file(cl_args.config_path().unwrap_or(Path::new(CONFIG_PATH))) {
         Ok(builder) => builder.apply_args(&cl_args).finish(),
         Err(e) => panic!("Failed to create the node config builder: {}", e),
     }
 }
 
-async fn start_entrynode<S: NodeStorageBackend>(config: NodeConfig<S>) {
-    let node_builder = EntryNodeBuilder::new(config.into());
+async fn start_entrynode<S: NodeStorageBackend>(local: Local, config: NodeConfig<S>) {
+    let entry_node_config = EntryNodeConfig::from(local, config);
+    let node_builder = EntryNodeBuilder::new(entry_node_config);
 
     match node_builder {
         Ok(builder) => match builder.finish().await {
@@ -75,11 +103,9 @@ async fn start_entrynode<S: NodeStorageBackend>(config: NodeConfig<S>) {
     }
 }
 
-async fn start_fullnode<S: NodeStorageBackend>(config: NodeConfig<S>)
-where
-    FullNodeConfig<Storage>: From<NodeConfig<S>>,
-{
-    let node_builder = FullNodeBuilder::<Storage>::new(config.into());
+async fn start_fullnode(local: Local, config: NodeConfig<Storage>) {
+    let full_node_config = FullNodeConfig::from(local, config);
+    let node_builder = FullNodeBuilder::<Storage>::new(full_node_config);
 
     match node_builder {
         Ok(builder) => match builder.with_plugin::<plugins::Mps>().finish().await {